+++ CHANGELOG.md
@@ -12,6 +12,7 @@ You can also check the [release page](https://github.com/visualize-admin/visuali
 - Features
   - It's now possible to animate the Map Chart 🎬 🗺️
   - It's now possible to normalize the data in Stacked Column Chart and Stacked Area Chart either in a static mode, or as an interactive filter 📊
+  - Normalized stacked charts, as well as Pie chart, now display both percentages and absolute values in the tooltip
   - Added support for a new dimension type - TemporalOrdinalDimension. This dimension can be used in all the places where you can use regular ordinal dimensions as well as to animate the chart
   - Scatterplot and Pie now animate updates
   - Axes now animate updates